+++ src/Logger.h
@@ -28,7 +28,7 @@ public:
     void logProcessingStart(int node_id, string error_code);
     void logFrameTransmission(int node_id, int seq_num,
                               const std::string &payload, char trailer, int modified,
-                              bool lost, int dup, int delay);
+                              bool lost, int dup, float delay);
     void logTimeoutEvent(int node_id, int seq_num);
     void logControlFrameSending(int node_id, bool ack_nack, int ack_num, bool loss);
     void logPayloadUploading(int seq_num, const std::string &payload);
